+++ changelog.md
@@ -2,6 +2,11 @@
 
 This project uses [Semantic Versioning](http://semver.org)
 
+## 0.8.1
+
+* When running suite, find the tests whose suite membership is declared in an
+  abstract class
+
 ## 0.8.0
 
 * Support the JUnit assumption API